+++ .gitignore
@@ -5,4 +5,4 @@ binance_secrets.json
 node_modules
 
 # Ignore output files
-*.txt
+outputs/**/*.txt
